Added Data Validation To The DTOs for validation upon requests

+++ Controllers/CommentController.cs
@@ -15,11 +15,13 @@ namespace InvestSense_API.Controllers
 		public readonly ApplicationDbContext _context;
 		public readonly IMapper _mapper;
 		public readonly ICommentRepository _commentRepository;
-		public CommentController(ApplicationDbContext context, IMapper mapper, ICommentRepository commentRepository)
+		public readonly IStockRepository _stockRepository;
+		public CommentController(ApplicationDbContext context, IMapper mapper, ICommentRepository commentRepository, IStockRepository stockRepository)
 		{
 			_context = context;
 			_mapper = mapper;
 			_commentRepository = commentRepository;
+			_stockRepository = stockRepository;
 		}
 
 		[HttpGet]
@@ -33,6 +35,10 @@ namespace InvestSense_API.Controllers
 		[HttpGet("{id}")]
 		public async Task<IActionResult> GetById([FromRoute] int id)
 		{
+			if (!ModelState.IsValid)
+			{
+				return BadRequest(ModelState);
+			}
 			var comment = await _commentRepository.GetByIdAsync(id);
 			if (comment == null)
 			{
@@ -42,27 +48,40 @@ namespace InvestSense_API.Controllers
 			return Ok(commentDTO);
 		}
 
-		[HttpPost]
-		public async Task<IActionResult> Create([FromBody] CreateCommentRequestDTO createdCommentDTO)
+		[HttpPost("{stockId:int}")]
+		public async Task<IActionResult> Create([FromRoute] int stockId, [FromBody] CreateCommentRequestDTO createdCommentDTO)
 		{
+			if (!ModelState.IsValid)
+			{
+				return BadRequest(ModelState);
+			}
+			if (!await _stockRepository.CheckStockExists(stockId))
+			{
+				return BadRequest("Stock Does Not Exist!");
+			}
 			var commentCreated = _mapper.Map<Comment>(createdCommentDTO);
+			commentCreated.StockId = stockId;
 			await _commentRepository.CreateAsync(commentCreated);
 			return CreatedAtAction(nameof(GetById), new { commentCreated.Id }, createdCommentDTO);
 		}
 
-		[HttpDelete("{id}")]
+		[HttpDelete("{id:int}")]
 		public async Task<IActionResult> Delete([FromRoute] int id)
 		{
+			if (!ModelState.IsValid)
+			{
+				return BadRequest(ModelState);
+			}
 			var existingComment = await _commentRepository.GetByIdAsync(id);
 			if (existingComment == null)
 			{
-				return NotFound();
+				return NotFound("Comment Does Not Exist!");
 			}
 
 			await _commentRepository.DeleteAsync(id);
 
 
-			return NoContent();
+			return Ok();
 		}
 	}
 }
